Better error message for missing tools
If the query planner tries to use a tool which doesn't exist, indicate
what went wrong.  Suggest that the user install the tool.

That doesn't work in some cases (since Linux users can't install mdfind,
for example).  I expect that this error will typically happen when the
user has configured his system to use some tool but that tool is
missing.