Use 'git grep' if possible
In most circumstances, when searching inside a Git repository, the
fastest thing you can possibly do is run 'git grep'.  It has some clever
tricks up its sleeve.

There are circumstances (like when 'mdfind' is available or when
restricting a search to files of a specific language) where we might be
able to outperform 'git grep', but we're not there yet.