Revert "Ignore files which git also ignores"
I'm not at all happy with this code.  Using 'git ls-files' requires a
second traversal of the entire repository.  For repositories with lots
of ignored files, this code also builds data structures that will never
be used. Ideally, I'd only consult gitignore files which apply to the
files we plan to search.

For example, if we're going to search these files, these are some of the
relevant gitignore files:

  foo/bar.txt : foo/.gitignore .gitignore
  baz.css : .gitignore

So we only need to consider two distinct gitignore files.  If the
repository has others, we don't care about them.

I hope that I can press 'git check-ignore' into service because the
ignore rules are very complex.  Trying to recreate them is bound to
cause hundreds of bugs.

This reverts commit 0ee84519bcdf95910281c3b691504040a2015618.