mdfind doesn't handle regular expressions
The mdfind utility can only search for literal strings.  If the user
provides a pattern that looks like a regular expression, have mdfind
search for only the literal prefix of that regex.  This still allows us
to use mdfind's index to quickly narrow the list of files to search, but
allows grep to perform the final regex matching.