Permanently disable the mdfind strategy
This strategy overlooks far too many legitimate search results.  For
example I just searched for "1000" in HTML files and it missed about 15
search results that it should have produced.

I'm not removing the code because I'd eventually like to enable it in
a few very specific circumstances where it's known to operate well.